requirements.txt
scikit-video
h5py==2.10.0
matplotlib==3.1.1
numpy==1.16.4
Pillow==6.0.0
scikit-image==0.16.1
scikit-learn==0.21.3
scipy==1.0.0
tensorboardX==1.8
torch==1.1.0
torchvision==0.3.0
visdom==0.1.8.8
